// src/types/csvRow.ts
export interface CsvRow {
  agent?: string;
  userType?: string;
  policy_number?: string;
  company_name?: string;
  category_name?: string;
  policy_start_date?: string;
  policy_end_date?: string;
  account_name?: string;
  email?: string;
  gender?: string;
  firstname?: string;
  phone?: string;
  address?: string;
  state?: string;
  zip?: string;
  dob?: string;
}